+++ packages/components/molecules/f-alert/jest.config.js
@@ -36,6 +36,7 @@ module.exports = {
 
     modulePathIgnorePatterns: [
         './test/component/',
-        './test/accessibility'
+        './test/accessibility',
+        './test/visual'
     ]
 };