+++ packages/components/molecules/f-mega-modal/stories/MegaModal.stories.js
@@ -1,69 +1,14 @@
-import { withKnobs, boolean, text } from '@storybook/addon-knobs';
 import { withA11y } from '@storybook/addon-a11y';
 import MegaModal from '../src/components/MegaModal.vue';
 
 export default {
     title: 'Components/Molecules',
-
-    decorators: [
-        withKnobs,
-        withA11y
-    ]
+    decorators: [withA11y]
 };
 
-export const MegaModalComponent = () => ({
+export const MegaModalComponent = (args, { argTypes }) => ({
     components: { MegaModal },
-
-    props: {
-        isOpen: {
-            default: boolean('Is open', true)
-        },
-
-        isNarrow: {
-            default: boolean('Is narrow', false)
-        },
-
-        isWide: {
-            default: boolean('Is wide', false)
-        },
-
-        isFlush: {
-            default: boolean('Is flush', false)
-        },
-
-        isFullHeight: {
-            default: boolean('Is full height', false)
-        },
-
-        isScrollable: {
-            default: boolean('Is scrollable', false)
-        },
-
-        isCloseFixed: {
-            default: boolean('Is close fixed', false)
-        },
-
-        isCloseRounded: {
-            default: boolean('Is close rounded', false)
-        },
-
-        hasOverlay: {
-            default: boolean('has overlay', true)
-        },
-
-        hasCloseButton: {
-            default: boolean('has close button', true)
-        },
-
-        closeOnBlur: {
-            default: boolean('Close on blur', true)
-        },
-
-        closeButtonCopy: {
-            default: text('Close button copy', 'Close modal')
-        }
-    },
-
+    props: Object.keys(argTypes),
     template: `
         <mega-modal
             :is-open="isOpen"
@@ -78,7 +23,7 @@ export const MegaModalComponent = () => ({
             :has-close-button="hasCloseButton"
             :close-on-blur="closeOnBlur"
             :close-button-copy="closeButtonCopy"
-            :title="'This place isn’t taking orders'">
+            :title="titleCopy">
 
             <p data-test-id="mega-modal-content">
                 Let's find another restaurant to order from.
@@ -87,5 +32,63 @@ export const MegaModalComponent = () => ({
         </mega-modal>
     `
 });
+
+MegaModalComponent.args = {
+    isOpen: true,
+    isNarrow: false,
+    isWide: false,
+    isFlush: false,
+    isFullHeight: false,
+    isScrollable: false,
+    isCloseFixed: false,
+    isCloseRounded: false,
+    hasOverlay: true,
+    hasCloseButton: true,
+    closeOnBlur: true,
+    closeButtonCopy: 'Close modal',
+    titleCopy: 'This place isn’t taking orders'
+};
+
+MegaModalComponent.argTypes = {
+    isOpen: {
+        control: { type: 'boolean' }
+    },
+    isNarrow: {
+        control: { type: 'boolean' }
+    },
+    isWide: {
+        control: { type: 'boolean' }
+    },
+    isFlush: {
+        control: { type: 'boolean' }
+    },
+    isFullHeight: {
+        control: { type: 'boolean' }
+    },
+    isScrollable: {
+        control: { type: 'boolean' }
+    },
+    isCloseFixed: {
+        control: { type: 'boolean' }
+    },
+    isCloseRounded: {
+        control: { type: 'boolean' }
+    },
+    hasOverlay: {
+        control: { type: 'boolean' }
+    },
+    hasCloseButton: {
+        control: { type: 'boolean' }
+    },
+    closeOnBlur: {
+        control: { type: 'boolean' }
+    },
+    closeButtonCopy: {
+        control: { type: 'text' }
+    },
+    titleCopy: {
+        control: { type: 'text' }
+    }
+};
 
 MegaModalComponent.storyName = 'f-mega-modal';