+++ packages/components/pages/f-checkout/stories/api/responses/getAddress.js
@@ -1,0 +1,25 @@
+import { Addresses } from '../../helpers/addresses';
+import { getSuccess, TENANTS } from '../../helpers';
+
+function getAddress (tenant) {
+    return {
+        ...Addresses[tenant]
+    };
+}
+
+function buildAddresses () {
+    const tenantRequests = [];
+
+    Object.values(TENANTS).forEach(tenant => {
+        tenantRequests.push({
+            url: `/address-${tenant}`,
+            ...getSuccess,
+            payload: getAddress(tenant)
+        });
+    });
+    return tenantRequests;
+}
+
+export default [
+    ...buildAddresses()
+];