+++ packages/components/pages/f-checkout/stories/api/responses/getCheckout/errors/index.js
@@ -1,0 +1,7 @@
+import forbidden from './forbidden';
+import server from './server';
+
+export default {
+    forbidden,
+    server
+};